move TrackSectionC
