+++ web/siteplan/src/feature/TrackDirectionFeature.ts
@@ -1,8 +1,9 @@
+import { TrackSectionC } from '@/enhanced_model/TrackSection'
 import LageplanFeature from '@/feature/LageplanFeature'
 import { SiteplanState } from '@/model/SiteplanModel'
 import { ISvgElement } from '@/model/SvgElement'
 import Track from '@/model/Track'
-import TrackSection, { TrackSectionC } from '@/model/TrackSection'
+import TrackSection from '@/model/TrackSection'
 import { distance, distanceCoords, normalizedDirection } from '@/util/Math'
 import SvgDraw from '@/util/SVG/Draw/SvgDraw'
 import { Feature } from 'ol'
